# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aqarly API",
    "description": "Backend for the Aqarly platform's apps.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/registrations": {
      "get": {
        "tags": [
          "registrations"
        ],
        "summary": "List Registrations",
        "description": "Waiting for a decision (oldest first), or every registration.",
        "operationId": "list_registrations_registrations_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "pending",
                "all"
              ],
              "type": "string",
              "default": "pending",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistrationOut"
                  },
                  "title": "Response List Registrations Registrations Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/registrations/{registration_id}/approve": {
      "post": {
        "tags": [
          "registrations"
        ],
        "summary": "Approve",
        "description": "Makes them a tenant and the unit's tenant, replacing whoever is there\nnow (`currentTenant`). They see their home the next time they load a page.",
        "operationId": "approve_registrations__registration_id__approve_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Registration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/registrations/{registration_id}/decline": {
      "post": {
        "tags": [
          "registrations"
        ],
        "summary": "Decline",
        "description": "They're told their building couldn't confirm them, and can register again.",
        "operationId": "decline_registrations__registration_id__decline_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Registration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "description": "Every refusal: a message the frontend can show as it is."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PropertyOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "address"
        ],
        "title": "PropertyOut"
      },
...
      "RegistrationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "unit": {
            "$ref": "#/components/schemas/UnitOut"
          },
          "property": {
            "$ref": "#/components/schemas/PropertyOut"
          },
          "currentTenant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TenantOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "approved",
                  "declined"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "decidedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decidedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "unit",
          "property",
          "currentTenant",
          "createdAt",
          "decision",
          "decidedAt"
        ],
        "title": "RegistrationOut"
      },
...
      "TenantOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "email"
        ],
        "title": "TenantOut"
      },
...
      "UnitOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "propertyId": {
            "type": "string",
            "title": "Propertyid"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "type": "string",
            "enum": [
              "occupied",
              "under-maintenance",
              "vacant"
            ],
            "title": "Status"
          },
          "tenantId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenantid"
          },
          "bedrooms": {
            "type": "integer",
            "title": "Bedrooms"
          },
          "bathrooms": {
            "type": "integer",
            "title": "Bathrooms"
          }
        },
        "type": "object",
        "required": [
          "id",
          "propertyId",
          "label",
          "status",
          "tenantId",
          "bedrooms",
          "bathrooms"
        ],
        "title": "UnitOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "The token from POST /auth/verify. In Swagger: Authorize, then paste it.",
        "in": "header",
        "name": "X-Session"
      }
    }
  }
}